+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### July 18, 2018 version 1.5.0
+ * KLV data added to output streams
+ * Output streams provide statistics on the # frames queued/sent
+ * Perform encoding on a separate thread for performance
+ * Attempt to use hardware-accelerated encoding when available
+ 
 ### July 11, 2018 version 1.4.0
  * Added Exec Maven plugin to allow running the viewer from the command line
  * Added command-line argument -DuberJar to produce an uber-JAR of jmisb-api